mcp_services/requirements.txt
fastmcp>=3.1.0
asyncpg                                 # PostgreSQL async driver (for MCP BMS)
opentelemetry-exporter-otlp-proto-grpc  # OTLP trace export to Tempo
